requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-richtextfield==1.6.1
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.1
pyparsing==2.4.7
PySocks==1.7.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.14
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
tornado==6.1
tzlocal==3.0
uritemplate==3.0.1
urllib3==1.26.6
